Count Odd Numbers below n
Update file "Count Odd Numbers below n.c"

+++ C/8 kyu/Count Odd Numbers below n.c
@@ -17,7 +17,7 @@ int main(void)
     return rc;
 }
 
-int input_number(int *n)
+int input_number(long *n)
 {
     int rc;
